requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1